Revert "enemyhealthbar"
This reverts commit 28f3c3f701aa5bd19c8f61839d79bd88eae71c91.

+++ Lock_And_Key/Assets/Scripts/Enemy&Player/EnemyHealth.cs
@@ -6,9 +6,9 @@ using UnityEngine.UI;
 public class EnemyHealth : MonoBehaviour
 {
 
-    public float maxHealth = 100;
+    public int maxHealth = 100;
     // public int maxHealth;
-    public float health;
+    public int health;
 
     public GameObject[] itemDrops;
